place turnRightToAligned with CmdShootTurnVision
replacing last factory so that all commands are defined in Commands folder and not spread out inline/factories/classes.

Change CmdInPlaceTurnVision to CmdShootTurnVision that does the in place turn that is interrupted by the vision aligning & shooting.

Not completely sure if the limelight turning on/off will work as intended